Added working nicknames function, fixed SearchBar component, deleted ChatContext componet
- Fixed SearchBar component - it was crashing application while trying to search for user that don’t exists
- Deleted ChatContext componet
- Case with setting nicknames :

In the end the best solution is:

Add nicknames (if they exists) to localstorage in updateChat function while cllicking on ContactItem in UserChats.

Get those nicknames from localstorage in ChaView instead of downloading them from firebase. (!!! In the end i had to keep that function to *instant update nickname in UserSettingCard and Header, also had to update nicknames in local storage every time user change nicknames in ChangeUserNicknamesCard component !!!)*

Forward ownerNickname and userNickname to Header and UserSettingsCard.